Removes get.js
